requirements.txt
fastapi-0.54.2
uvicorn
async-generator
pymongo
dnszone